Screen now has two parts for 3DS development
The intention all along with this project has been to make a 3DS game,
and I decided that even if it significantly sacrifices the usability of
the PC version, I should be working with two screens. Thus, love.draw()
now renders two virtual screens, then displays each screen separately.
The bottom screen is given more screen real-estate, because the plan is
for cards to be managed entirely using the touchscreen. The top screen
will be little more than a place to put a scoreboard, round counter,
etc.